+++ php_rpminfo.h
@@ -2,17 +2,13 @@
   +----------------------------------------------------------------------+
   | rpminfo extension for PHP                                            |
   +----------------------------------------------------------------------+
-  | Copyright (c) The PHP Group                                          |
+  | SPDX-FileCopyrightText: Copyright (c) Remi Collet <remi@php.net>     |
   +----------------------------------------------------------------------+
-  | This source file is subject to version 3.01 of the PHP license,      |
-  | that is bundled with this package in the file LICENSE, and is        |
-  | available through the world-wide-web at the following url:           |
-  | http://www.php.net/license/3_01.txt.                                 |
-  | If you did not receive a copy of the PHP license and are unable to   |
-  | obtain it through the world-wide-web, please send a note to          |
-  | license@php.net so we can mail you a copy immediately.               |
-  +----------------------------------------------------------------------+
-  | Author: Remi Collet <remi@php.net>                                   |
+  | This source file is subject to the Modified BSD License that is      |
+  | bundled with this package in the file LICENSE, and is available      |
+  | through the WWW at <https://opensource.org/license/BSD-3-Clause>.    |
+  |                                                                      |
+  | SPDX-License-Identifier: BSD-3-Clause                                |
   +----------------------------------------------------------------------+
 */
 
@@ -24,7 +20,7 @@ extern zend_module_entry rpminfo_module_entry;
 
 #define PHP_RPMINFO_VERSION "1.2.2-dev"
 #define PHP_RPMINFO_AUTHOR  "Remi Collet"
-#define PHP_RPMINFO_LICENSE "PHP-3.01"
+#define PHP_RPMINFO_LICENSE "BSD-3-Clause"
 
 #ifdef ZTS
 #include "TSRM.h"