+++ rpminfo_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: dc980a56084190700162f5a8c70b54fcdbc30ced */
+ * Stub hash: 5ae46dc3cb77804d128641a8032d01723c369066 */
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_rpmaddtag, 0, 1, _IS_BOOL, 0)
 	ZEND_ARG_TYPE_INFO(0, rpmtag, IS_LONG, 0)
@@ -46,7 +46,6 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_rpmdefine, 0, 1, _IS_BOOL, 0)
 	ZEND_ARG_TYPE_INFO(0, macro, IS_STRING, 0)
 ZEND_END_ARG_INFO()
 
-
 ZEND_FUNCTION(rpmaddtag);
 ZEND_FUNCTION(rpmdbinfo);
 ZEND_FUNCTION(rpmdbsearch);
@@ -56,7 +55,6 @@ ZEND_FUNCTION(rpmgetsymlink);
 ZEND_FUNCTION(rpmexpand);
 ZEND_FUNCTION(rpmexpandnumeric);
 ZEND_FUNCTION(rpmdefine);
-
 
 static const zend_function_entry ext_functions[] = {
 	ZEND_FE(rpmaddtag, arginfo_rpmaddtag)